Alpha - Sprint 1 Review
Revamp player controller. Now we use the character's Rigidbody to calculate it's physiscs, instead of having movement being calculated mathematecally and walljump/grappling hook with physics.
Implement spike block. This block is triggered when the player passes under it, causing it to get an impulse to fall. After hitting either the player or the ground.
It's not yet implemented the resetting of these blocks that are destroyed when you change rooms.
Implement camera transitions. Camera is centered in each room, and only moves when the player crosses a trigger from one virtual cam to another, causing the main cam to move to the next center of the room. In this build the camera mves a bit to the left and right, it was a mistake while doing the collider be the size of the camera viewport.

This is the start of a new state of the game. By the end of it we look foward to having all gameplay aspects implemented, as well as the levels and some enemies.