Respawn on destroyable objects
Code respawn timer on objects that are destroyed due to player action (falling platforms and spikes).
Each time the object respawns, it instantiates a new prefab of itself, having a fixed position has a reference.
Rename some of the prefabs, there is still some "general" cleaning needed in the project.